# apps/web/openapi.json
{"openapi":"3.1.0","info":{"title":"Brier API","description":"AI Agent 工作空间平台","license":{"name":""},"version":"0.1.0"},"paths":{"/api/auth/logout":{"get":{"tags":["crate::routes::auth"],"summary":"登出：清除会话 Cookie 并跳转首页。","operationId":"logout","responses":{"302":{"description":"清除 Cookie 并跳转"}}}},"/api/auth/me":{"get":{"tags":["crate::routes::auth"],"summary":"返回当前登录用户信息。","operationId":"auth_me","responses":{"200":{"description":"当前用户信息","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"未登录"}}}},"/api/auth/{provider}/callback":{"get":{"tags":["crate::routes::auth"],"summary":"OAuth 回调：换码 → 取身份 → 查找/创建账户 → 签发会话 Cookie 并跳转前端。","operationId":"provider_callback","parameters":[{"name":"provider","in":"path","description":"OAuth 提供方，如 github / gitee","required":true,"schema":{"type":"string"}},{"name":"code","in":"query","description":"OAuth 授权码（OAuth 回调携带的 query 参数）。","required":true,"schema":{"type":"string"}}],"responses":{"302":{"description":"登录成功，Set-Cookie 会话令牌并跳转"},"401":{"description":"换码/取身份失败"},"404":{"description":"未知 provider"}}}},"/api/auth/{provider}/login":{"get":{"tags":["crate::routes::auth"],"summary":"跳转指定 OAuth 提供方的授权页；未注册的 provider 返回 404。","operationId":"provider_login","parameters":[{"name":"provider","in":"path","description":"OAuth 提供方，如 github / gitee","required":true,"schema":{"type":"string"}}],"responses":{"302":{"description":"跳转到授权页"},"404":{"description":"未知 provider"}}}},"/api/work-computers":{"get":{"tags":["crate::routes::work_computer"],"operationId":"list_work_computers","responses":{"200":{"description":"当前用户的工作电脑列表","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkComputer"}}}}},"401":{"description":"未登录"}}},"post":{"tags":["crate::routes::work_computer"],"operationId":"create_work_computer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkComputerRequest"}}},"required":true},"responses":{"200":{"description":"创建成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkComputer"}}}},"400":{"description":"请求体不合法"},"401":{"description":"未登录"}}}},"/api/work-computers/{computer_id}":{"get":{"tags":["crate::routes::work_computer"],"operationId":"get_work_computer","parameters":[{"name":"computer_id","in":"path","description":"工作电脑 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkComputerId"}}],"responses":{"200":{"description":"工作电脑详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkComputer"}}}},"401":{"description":"未登录"},"404":{"description":"工作电脑不存在"}}}},"/api/workspaces":{"get":{"tags":["crate::routes::workspace"],"summary":"列出当前用户的全部空间（需登录）。","operationId":"list_workspaces","responses":{"200":{"description":"空间列表","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Workspace"}}}}},"401":{"description":"未登录"}}},"post":{"tags":["crate::routes::workspace"],"summary":"创建空间（需登录）。","operationId":"create_workspace","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkspaceRequest"}}},"required":true},"responses":{"200":{"description":"创建成功，返回空间","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Workspace"}}}},"400":{"description":"请求体不合法"},"401":{"description":"未登录"}}}},"/api/workspaces/{workspace_id}":{"get":{"tags":["crate::routes::workspace"],"summary":"获取工作空间详情（需登录 + 空间成员或创建者）。","operationId":"get_workspace","parameters":[{"name":"workspace_id","in":"path","description":"工作空间 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceId"}}],"responses":{"200":{"description":"工作空间详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Workspace"}}}},"401":{"description":"未登录"},"404":{"description":"空间不存在或无权限"}}}},"/api/workspaces/{workspace_id}/agents":{"get":{"tags":["crate::routes::agent"],"operationId":"list_agents","parameters":[{"name":"workspace_id","in":"path","description":"工作空间 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceId"}}],"responses":{"200":{"description":"Agent 列表","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Agent"}}}}},"401":{"description":"未登录"},"404":{"description":"空间不存在或无权限"}}},"post":{"tags":["crate::routes::agent"],"operationId":"create_agent","parameters":[{"name":"workspace_id","in":"path","description":"工作空间 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceId"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAgentRequest"}}},"required":true},"responses":{"200":{"description":"创建成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Agent"}}}},"401":{"description":"未登录"},"404":{"description":"空间不存在或无权限"}}}},"/api/workspaces/{workspace_id}/agents/{agent_id}":{"get":{"tags":["crate::routes::agent"],"operationId":"get_agent","parameters":[{"name":"workspace_id","in":"path","description":"工作空间 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceId"}},{"name":"agent_id","in":"path","description":"Agent ID","required":true,"schema":{"$ref":"#/components/schemas/AgentId"}}],"responses":{"200":{"description":"Agent 详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Agent"}}}},"401":{"description":"未登录"},"404":{"description":"Agent 不存在或无权限"}}},"delete":{"tags":["crate::routes::agent"],"operationId":"delete_agent","parameters":[{"name":"workspace_id","in":"path","description":"工作空间 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceId"}},{"name":"agent_id","in":"path","description":"Agent ID","required":true,"schema":{"$ref":"#/components/schemas/AgentId"}}],"responses":{"200":{"description":"删除成功"},"401":{"description":"未登录"},"404":{"description":"Agent 不存在或无权限"}}}},"/api/workspaces/{workspace_id}/teams":{"get":{"tags":["crate::routes::team"],"operationId":"list_teams","parameters":[{"name":"workspace_id","in":"path","description":"工作空间 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceId"}}],"responses":{"200":{"description":"Team 列表","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AgentTeam"}}}}},"401":{"description":"未登录"},"404":{"description":"空间不存在或无权限"}}},"post":{"tags":["crate::routes::team"],"operationId":"create_team","parameters":[{"name":"workspace_id","in":"path","description":"工作空间 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceId"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTeamRequest"}}},"required":true},"responses":{"200":{"description":"创建成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentTeam"}}}},"401":{"description":"未登录"},"404":{"description":"空间不存在或无权限"}}}},"/api/workspaces/{workspace_id}/teams/{team_id}":{"get":{"tags":["crate::routes::team"],"operationId":"get_team","parameters":[{"name":"workspace_id","in":"path","description":"工作空间 ID","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceId"}},{"name":"team_id","in":"path","description":"Team ID","required":true,"schema":{"$ref":"#/components/schemas/AgentTeamId"}}],"responses":{"200":{"description":"Team 详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentTeam"}}}},"401":{"description":"未登录"},"404":{"description":"Team 不存在或无权限"}}}},"/api/{provider}/repos":{"get":{"tags":["crate::routes::forge"],"summary":"列出指定 provider 下当前用户的仓库（需登录 + 已绑定该 provider 令牌）。\n未注册的 provider 返回 404；provider 有配置但未绑定令牌返回 Auth 错误。","operationId":"list_repos","parameters":[{"name":"provider","in":"path","description":"代码托管平台名，如 github / gitee","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"仓库列表","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RepoInfo"}}}}},"401":{"description":"未登录或未绑定令牌"},"404":{"description":"未知 provider"}}}}},"components":{"schemas":{"Agent":{"type":"object","required":["id","workspace_id","creator_id","name","status","visibility","created_at","updated_at"],"properties":{"color":{"type":["string","null"]},"created_at":{"type":"string","format":"date-time"},"creator_id":{"$ref":"#/components/schemas/UserId"},"description":{"type":["string","null"]},"icon":{"type":["string","null"]},"id":{"$ref":"#/components/schemas/AgentId"},"last_active":{"type":["string","null"],"format":"date-time"},"name":{"type":"string"},"public_scope":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/PublicScope"}]},"runtime":{"type":["string","null"]},"status":{"$ref":"#/components/schemas/AgentStatus"},"updated_at":{"type":"string","format":"date-time"},"visibility":{"$ref":"#/components/schemas/AgentVisibility"},"work_computer_id":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/WorkComputerId"}]},"workspace_id":{"$ref":"#/components/schemas/WorkspaceId"}}},"AgentId":{"type":"string","format":"uuid","description":"Agent 唯一标识"},"AgentStatus":{"type":"string","enum":["online","offline","connecting"]},"AgentTeam":{"type":"object","required":["id","workspace_id","creator_id","name","mode","status","created_at","updated_at"],"properties":{"created_at":{"type":"string","format":"date-time"},"creator_id":{"$ref":"#/components/schemas/UserId"},"description":{"type":["string","null"]},"id":{"$ref":"#/components/schemas/AgentTeamId"},"mode":{"$ref":"#/components/schemas/TeamMode"},"name":{"type":"string"},"primary_agent_id":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/AgentId"}]},"status":{"$ref":"#/components/schemas/TeamStatus"},"updated_at":{"type":"string","format":"date-time"},"workspace_id":{"$ref":"#/components/schemas/WorkspaceId"}}},"AgentTeamId":{"type":"string","format":"uuid","description":"Agent 团队唯一标识"},"AgentVisibility":{"type":"string","enum":["private","public"]},"CreateAgentRequest":{"type":"object","required":["name"],"properties":{"color":{"type":["string","null"]},"description":{"type":["string","null"]},"icon":{"type":["string","null"]},"name":{"type":"string"},"public_scope":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/PublicScope"}]},"runtime":{"type":["string","null"]},"visibility":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/AgentVisibility"}]},"work_computer_id":{"type":["string","null"]}}},"CreateTeamRequest":{"type":"object","required":["name"],"properties":{"description":{"type":["string","null"]},"mode":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/TeamMode"}]},"name":{"type":"string"},"primary_agent_id":{"type":["string","null"]}}},"CreateWorkComputerRequest":{"type":"object","required":["name","computer_type","host","os"],"properties":{"computer_type":{"$ref":"#/components/schemas/WorkComputerType"},"host":{"type":"string"},"name":{"type":"string"},"os":{"type":"string"}}},"CreateWorkspaceRequest":{"type":"object","required":["name","slug"],"properties":{"auto_issue_assign":{"type":["boolean","null"]},"auto_pr_review":{"type":["boolean","null"]},"avatar":{"type":["string","null"]},"description":{"type":["string","null"]},"instructions":{"type":["string","null"]},"name":{"type":"string"},"repositories":{"type":["array","null"],"items":{"type":"string"}},"slug":{"type":"string"}}},"PublicScope":{"type":"string","enum":["all","joined_spaces","specified_spaces"]},"RepoInfo":{"type":"object","description":"代码仓库信息（厂商 API 返回的仓库视图，非 Brier 领域概念）。\n与 `RepositoryProvider` 强绑定（trait 方法签名引用），故随契约同居本包；\n消费方仅限适配层（brier-forge）与应用层（brier-api）两端，不进 brier-type。","required":["full_name","name","private"],"properties":{"full_name":{"type":"string"},"name":{"type":"string"},"private":{"type":"boolean"}}},"TeamMode":{"type":"string","enum":["coordinator","sequential","graph"]},"TeamStatus":{"type":"string","enum":["available","unavailable"]},"User":{"type":"object","description":"账户本体。第三方身份（OAuth）与本地凭证（手机号/密码）统一挂载在\n`id`（用户 UUID）之下；password_hash 属敏感凭证，永不进入该 API 模型。","required":["id","username","status","created_at","updated_at"],"properties":{"avatar_url":{"type":["string","null"]},"created_at":{"type":"string","format":"date-time"},"email":{"type":["string","null"]},"id":{"$ref":"#/components/schemas/UserId"},"last_login_at":{"type":["string","null"],"format":"date-time"},"name":{"type":["string","null"]},"phone":{"type":["string","null"]},"status":{"$ref":"#/components/schemas/UserStatus"},"updated_at":{"type":"string","format":"date-time"},"username":{"type":"string"}}},"UserId":{"type":"string","format":"uuid","description":"用户唯一标识"},"UserStatus":{"type":"string","enum":["active","disabled"]},"WorkComputer":{"type":"object","required":["id","user_id","name","computer_type","host","os","status","created_at","updated_at"],"properties":{"computer_type":{"$ref":"#/components/schemas/WorkComputerType"},"created_at":{"type":"string","format":"date-time"},"host":{"type":"string"},"id":{"$ref":"#/components/schemas/WorkComputerId"},"name":{"type":"string"},"os":{"type":"string"},"status":{"$ref":"#/components/schemas/WorkComputerStatus"},"updated_at":{"type":"string","format":"date-time"},"user_id":{"$ref":"#/components/schemas/UserId"}}},"WorkComputerId":{"type":"string","format":"uuid","description":"工作电脑唯一标识"},"WorkComputerStatus":{"type":"string","enum":["online","offline"]},"WorkComputerType":{"type":"string","enum":["local","ssh","cloud"]},"Workspace":{"type":"object","required":["id","creator_id","name","slug","repositories","auto_pr_review","auto_issue_assign","created_at","updated_at"],"properties":{"auto_issue_assign":{"type":"boolean"},"auto_pr_review":{"type":"boolean"},"avatar":{"type":["string","null"]},"created_at":{"type":"string","format":"date-time"},"creator_id":{"$ref":"#/components/schemas/UserId"},"description":{"type":["string","null"]},"id":{"$ref":"#/components/schemas/WorkspaceId"},"instructions":{"type":["string","null"]},"name":{"type":"string"},"repositories":{"type":"array","items":{"type":"string"}},"slug":{"type":"string"},"updated_at":{"type":"string","format":"date-time"}}},"WorkspaceId":{"type":"string","format":"uuid","description":"工作空间唯一标识"}}}}